+++ addOns/openapi/src/test/java/org/zaproxy/zap/extension/openapi/ExtensionOpenApiTest.java
@@ -39,22 +39,12 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 import org.apache.commons.httpclient.URI;
 import org.apache.commons.httpclient.URIException;
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.db.RecordHistory;
@@ -65,8 +55,6 @@ import org.parosproxy.paros.model.HistoryReference;
 import org.parosproxy.paros.model.Model;
 import org.zaproxy.zap.extension.ascan.VariantFactory;
 import org.zaproxy.zap.extension.spider.ExtensionSpider;
-import org.zaproxy.zap.model.Context;
-import org.zaproxy.zap.model.StructuralNodeModifier;
 import org.zaproxy.zap.testutils.NanoServerHandler;
 import org.zaproxy.zap.utils.I18N;
 import org.zaproxy.zap.utils.ZapXmlConfiguration;
@@ -193,46 +181,6 @@ class ExtensionOpenApiTest extends AbstractServerTest {
         assertThat(results.getHistoryReferences(), hasSize(1));
     }
 
-    @ParameterizedTest
-    @NullAndEmptySource
-    void shouldGenerateDataDrivenNodesOnContext(String target) throws IOException {
-        // Given
-        File file = createLocalDefinition("v3/PetStore_defn.json").toFile();
-        Context ctx = createContext();
-        String serverUrl = "http://localhost:" + nano.getListeningPort();
-        String targetUrl = target != null ? serverUrl + "/v1" : null;
-        String expectedUrl = target != null ? targetUrl : serverUrl + "/PetStore";
-
-        // When
-        extensionOpenApi.importOpenApiDefinition(file, targetUrl, false, ctx.getId());
-
-        // Then
-        assertThat(
-                ctx.getDataDrivenNodes(),
-                contains(
-                        expectedUrl + "(/pet/)(.+?)(/.*)",
-                        expectedUrl + "(/store/order/)(.+?)(/.*)",
-                        expectedUrl + "(/user/)(.+?)(/.*)"));
-    }
-
-    @Test
-    void shouldGenerateDataDrivenNodesOnContextForMultiVarPath() throws IOException {
-        // Given
-        File file = createLocalDefinition("v3/MultiVarPath_defn.yaml").toFile();
-        Context ctx = createContext();
-        String serverUrl = "http://localhost:" + nano.getListeningPort();
-
-        // When
-        extensionOpenApi.importOpenApiDefinition(file, null, false, ctx.getId());
-
-        // Then
-        assertThat(
-                ctx.getDataDrivenNodes(),
-                contains(
-                        serverUrl + "(/api/stuff/.+?/subthing/)(.+?)(/.*)",
-                        serverUrl + "(/api/stuff/)(.+?)(/.*)"));
-    }
-
     private Path createLocalDefinition(String path) throws IOException {
         Path directory = Files.createTempDirectory("local-defn");
         Path localDefinition = directory.resolve(path.substring(path.lastIndexOf("/") + 1));
@@ -257,10 +205,6 @@ class ExtensionOpenApiTest extends AbstractServerTest {
         return localDefinition.toFile();
     }
 
-    private static Context createContext() {
-        return Model.getSingleton().getSession().getNewContext("Test Context");
-    }
-
     private static class EmptyServerHandler extends NanoServerHandler {
 
         EmptyServerHandler() {
@@ -271,56 +215,5 @@ class ExtensionOpenApiTest extends AbstractServerTest {
         protected NanoHTTPD.Response serve(NanoHTTPD.IHTTPSession session) {
             return newFixedLengthResponse("");
         }
-    }
-
-    private static Matcher<List<StructuralNodeModifier>> contains(String... regexes) {
-        List<String> expectedValues = new ArrayList<>();
-        expectedValues.addAll(Arrays.asList(regexes));
-        Collections.sort(expectedValues);
-
-        return new BaseMatcher<List<StructuralNodeModifier>>() {
-
-            @Override
-            public boolean matches(Object actualValue) {
-                @SuppressWarnings("unchecked")
-                List<StructuralNodeModifier> values = (List<StructuralNodeModifier>) actualValue;
-                if (values.isEmpty()) {
-                    return false;
-                }
-
-                List<String> matched = new ArrayList<>(expectedValues);
-                for (StructuralNodeModifier value : values) {
-                    if (!matched.remove(value.getPattern().pattern())) {
-                        return false;
-                    }
-                }
-                return matched.isEmpty();
-            }
-
-            @Override
-            public void describeTo(Description description) {
-                description
-                        .appendText("the DDN regular expressions to be ")
-                        .appendValue(expectedValues);
-            }
-
-            @Override
-            public void describeMismatch(Object item, Description description) {
-                @SuppressWarnings("unchecked")
-                List<StructuralNodeModifier> values = (List<StructuralNodeModifier>) item;
-                if (values.isEmpty()) {
-                    description.appendText("had no DDNs");
-                } else {
-                    description
-                            .appendText("were ")
-                            .appendValue(
-                                    values.stream()
-                                            .map(StructuralNodeModifier::getPattern)
-                                            .map(Pattern::pattern)
-                                            .sorted()
-                                            .collect(Collectors.toList()));
-                }
-            }
-        };
     }
 }